Day17: add optimisations

+++ Day17/io.cpp
@@ -6,11 +6,11 @@
 
 void day17::Input17::set_content(std::vector<std::string> lines)
 {
-    std::vector<std::vector<int>> int_map(lines.size());
+    std::vector<std::vector<uint8_t>> int_map(lines.size());
 
     for (int i=0; i<lines.size(); ++i) {
         int_map[i].resize(lines[i].size());
-        for (int j=0; j<lines[i].size(); ++j) {
+        for (uint8_t j=0; j<lines[i].size(); ++j) {
             int_map[i][j] = lines[i][j] - '0';
         }
